Refactors setup logic out of the main server file.
