+++ server/config/session.js
@@ -1,0 +1,8 @@
+const sessionConfig = {
+  secret: 'dont-forget-that-secret',
+  cookie: {},
+  resave: false,
+  saveUninitialized: true
+};
+
+module.exports = sessionConfig;
